Exception handling - delete

+++ src/main/java/com/vonvit/course/services/UserService.java
@@ -2,8 +2,10 @@ package com.vonvit.course.services;
 
 import com.vonvit.course.entities.User;
 import com.vonvit.course.repositories.UserRepository;
+import com.vonvit.course.services.exceptions.DatabaseException;
 import com.vonvit.course.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Repository;
 import org.springframework.stereotype.Service;
 
@@ -30,7 +32,16 @@ public class UserService {
     }
 
     public void delete(Long id){
-        repository.deleteById(id);
+        try {
+            if(!repository.existsById(id)) throw new ResourceNotFoundException(id);
+            repository.deleteById(id);
+        }
+        catch (ResourceNotFoundException e){
+            throw new ResourceNotFoundException(id);
+        } catch (DataIntegrityViolationException e){
+            throw new DatabaseException(e.getMessage());
+        }
+
     }
 
     public User update(Long id, User obj){
